chore: type storybook constants
